migration: Recreate InitialCreate migration

+++ src/MarketStat.Common/MarketStat.Common.Converter/MarketStat.Common.Converter.Dimensions/DimDateConverter.cs
@@ -14,9 +14,9 @@ public class DimDateConverter
         {
             DateId = dimDate.DateId,
             FullDate = dimDate.FullDate,
-            Year = (short)dimDate.Year,
-            Quarter = (short)dimDate.Quarter,
-            Month = (short)dimDate.Month
+            Year = dimDate.Year,
+            Quarter = dimDate.Quarter,
+            Month = dimDate.Month
         };
     }
